fix _ 싱글톤 템플릿화 , 사용되지 않는 클래스 제거, 클라이언트 개발, 데이터베이스 64->32bit 전환 # 세부사항 1. 싱글톤을 템플릿화 하면서 동시에 필요없는 클래스들(engine,server관련 mgr) 제거 2. mysql 64bit -> 32bit 전환과 동시에 int64_t 형식 전부 int형으로 변경 3. 클라이언트 일부 기능 개발

+++ KaKaoProgram/Src/TCPSocket.cpp
@@ -3,7 +3,7 @@
 
 int TCPSocket::Send(const char* buf, const int bufSize)
 {
-	int sendSize = send(m_Socket, buf, bufSize, 0);
+	int sendSize = send(m_Socket, buf, static_cast<int>(bufSize), 0);
 
 	// WOULD BLOCK과 연결종료 같은것들은 외부에서 처리하자.
 	if (sendSize < 0)
